refactor certificate template form into reusable component
